Refactor: move channel relaying to a ChannelRelayer actor. The relayer is the top-level orchestrator that forwards to a payment handler, channel relayer or (coming) node relayer.
