+++ eclair-core/src/main/scala/fr/acinq/eclair/Setup.scala
@@ -45,7 +45,8 @@ import fr.acinq.eclair.db.{BackupHandler, Databases}
 import fr.acinq.eclair.io.{Authenticator, Server, Switchboard}
 import fr.acinq.eclair.payment.receive.PaymentHandler
 import fr.acinq.eclair.payment.send.{Autoprobe, PaymentInitiator}
-import fr.acinq.eclair.payment.{Auditor, Relayer}
+import fr.acinq.eclair.payment.Auditor
+import fr.acinq.eclair.payment.relay.Relayer
 import fr.acinq.eclair.router._
 import fr.acinq.eclair.tor.TorProtocolHandler.OnionServiceVersion
 import fr.acinq.eclair.tor.{Controller, TorProtocolHandler}
