+++ eclair-core/src/test/scala/fr/acinq/eclair/channel/ThroughputSpec.scala
@@ -27,7 +27,7 @@ import fr.acinq.eclair.TestConstants.{Alice, Bob}
 import fr.acinq.eclair._
 import fr.acinq.eclair.blockchain._
 import fr.acinq.eclair.blockchain.bitcoind.ZmqWatcher
-import fr.acinq.eclair.payment.Relayer
+import fr.acinq.eclair.payment.relay.Relayer
 import fr.acinq.eclair.wire.{Init, UpdateAddHtlc}
 import org.scalatest.FunSuite
